added new util for getServerFeatures
